refactor: split electron main process
