+++ electron/main.js
@@ -1,8 +1,9 @@
-const { app, BrowserWindow, shell, Menu, ipcMain, dialog, safeStorage, session } = require('electron')
+const { app, BrowserWindow, shell, Menu } = require('electron')
 const fs = require('fs')
 const path = require('path')
-const { spawn } = require('child_process')
 const { createAutoUpdateService } = require('./autoUpdater')
+const { createBackendProcessManager } = require('./backendProcess')
+const { registerIpcHandlers } = require('./ipcHandlers')
 
 // 禁用开发时的 CSP 警告
 if (process.env.NODE_ENV === 'development' || !app.isPackaged) {
@@ -10,7 +11,6 @@ if (process.env.NODE_ENV === 'development' || !app.isPackaged) {
 }
 
 let mainWindow
-let backendProcess
 
 function createWindow() {
   // 隐藏菜单栏
@@ -78,87 +78,21 @@ function createWindow() {
   mainWindow.on('closed', () => {
     mainWindow = null
   })
-}
 
-function startBackend() {
-  const fs = require('fs')
-  let backendPath
-
-  if (!app.isPackaged) {
-    // 开发环境：使用 backend 目录下编译好的可执行文件
-    const exeName = process.platform === 'win32' ? 'knot-backend.exe' : 'knot-backend'
-    backendPath = path.join(__dirname, '../backend', exeName)
-
-    // 如果开发环境下可执行文件不存在，提示需要先编译
-    if (!fs.existsSync(backendPath)) {
-      console.error('错误: 后端可执行文件不存在:', backendPath)
-      console.error('请先在 backend 目录下运行: go build -o ' + exeName)
-      return
-    }
-
-    console.log('开发模式 - 正在启动后端:', backendPath)
-  } else {
-    // 生产环境：使用打包后的可执行文件
-    const resourcesPath = process.resourcesPath
-    const exeName = process.platform === 'win32' ? 'knot-backend.exe' : 'knot-backend'
-    backendPath = path.join(resourcesPath, 'bin', exeName)
-
-    console.log('生产模式 - 资源目录:', resourcesPath)
-    console.log('生产模式 - 后端路径:', backendPath)
-    console.log('生产模式 - 文件是否存在:', fs.existsSync(backendPath))
-
-    if (!fs.existsSync(backendPath)) {
-      console.error('错误: 后端可执行文件不存在:', backendPath)
-      return
-    }
-
-    // Linux/macOS: 确保后端有执行权限
-    if (process.platform !== 'win32') {
-      try {
-        const stats = fs.statSync(backendPath)
-        const mode = stats.mode
-        if ((mode & 0o111) === 0) {
-          console.log('修复后端执行权限...')
-          fs.chmodSync(backendPath, 0o755)
-        }
-      } catch (err) {
-        console.error('检查/设置权限失败:', err)
-      }
-    }
-  }
-
-  backendProcess = spawn(backendPath, [], {
-    stdio: ['pipe', 'pipe', 'pipe']
+  mainWindow.on('maximize', () => {
+    mainWindow.webContents.send('window-maximized-state', true)
   })
-
-  backendProcess.stdout.on('data', (data) => {
-    console.log('后端输出:', data.toString())
-  })
-
-  backendProcess.stderr.on('data', (data) => {
-    console.error('后端错误/日志:', data.toString())
-  })
-
-  backendProcess.on('error', (err) => {
-    console.error('启动后端失败:', err)
-  })
-
-  backendProcess.on('exit', (code) => {
-    console.log(`后端退出，退出码: ${code}`)
+  mainWindow.on('unmaximize', () => {
+    mainWindow.webContents.send('window-maximized-state', false)
   })
 }
 
-function stopBackend() {
-  if (backendProcess) {
-    backendProcess.kill()
-    backendProcess = null
-  }
-}
+const backendManager = createBackendProcessManager({ app, appDir: __dirname })
 
 const updateService = createAutoUpdateService({
   app,
   getMainWindow: () => mainWindow,
-  stopBackend
+  stopBackend: backendManager.stop
 })
 
 // 单实例锁
@@ -175,151 +109,13 @@ if (!gotTheLock) {
   })
 
   app.whenReady().then(() => {
-    ipcMain.handle('select-folder', async () => {
-      const result = await dialog.showOpenDialog(mainWindow, {
-        properties: ['openDirectory'],
-        title: '选择文件夹'
-      })
-      if (!result.canceled && result.filePaths.length > 0) {
-        return result.filePaths[0]
-      }
-      return null
+    registerIpcHandlers({
+      app,
+      getMainWindow: () => mainWindow,
+      updateService
     })
 
-    ipcMain.handle('open-folder', async (event, folderPath) => {
-      if (!folderPath) return false
-      try {
-        const error = await shell.openPath(folderPath)
-        return error === ''
-      } catch (err) {
-        console.error('打开目录失败:', err)
-        return false
-      }
-    })
-
-    ipcMain.handle('open-external', async (event, url) => {
-      if (!url || !/^https?:\/\//i.test(url)) return false
-      try {
-        await shell.openExternal(url)
-        return true
-      } catch (err) {
-        console.error('打开外部链接失败:', err)
-        return false
-      }
-    })
-
-    ipcMain.handle('encrypt-password', async (event, password) => {
-      if (!password) return null
-      try {
-        if (safeStorage.isEncryptionAvailable()) {
-          const encrypted = safeStorage.encryptString(password)
-          return encrypted.toString('base64')
-        } else {
-          console.warn('safeStorage 加密不可用，使用 base64 编码')
-          return Buffer.from(password).toString('base64')
-        }
-      } catch (err) {
-        console.error('加密失败:', err)
-        return null
-      }
-    })
-
-    ipcMain.handle('decrypt-password', async (event, encryptedBase64) => {
-      if (!encryptedBase64) return null
-      try {
-        if (safeStorage.isEncryptionAvailable()) {
-          const encrypted = Buffer.from(encryptedBase64, 'base64')
-          return safeStorage.decryptString(encrypted)
-        } else {
-          return Buffer.from(encryptedBase64, 'base64').toString('utf-8')
-        }
-      } catch (err) {
-        console.error('解密失败:', err)
-        return null
-      }
-    })
-
-    ipcMain.handle('is-encryption-available', async () => {
-      return safeStorage.isEncryptionAvailable()
-    })
-
-    ipcMain.on('get-app-version', (event) => {
-      event.returnValue = app.getVersion()
-    })
-
-    ipcMain.handle('update-get-status', () => {
-      return updateService.getStatus()
-    })
-
-    ipcMain.handle('update-check', async () => {
-      return updateService.checkForUpdates()
-    })
-
-    ipcMain.handle('update-download', async () => {
-      return updateService.downloadUpdate()
-    })
-
-    ipcMain.handle('update-install', () => {
-      return updateService.installUpdate()
-    })
-
-    // 窗口控制事件
-    ipcMain.handle('window-minimize', () => {
-      if (mainWindow) mainWindow.minimize()
-    })
-
-    ipcMain.handle('window-maximize', () => {
-      if (mainWindow) {
-        if (mainWindow.isMaximized()) {
-          mainWindow.unmaximize()
-        } else {
-          mainWindow.maximize()
-        }
-      }
-    })
-
-    ipcMain.handle('window-close', () => {
-      if (mainWindow) mainWindow.close()
-    })
-
-    ipcMain.handle('window-is-maximized', () => {
-      return mainWindow ? mainWindow.isMaximized() : false
-    })
-
-    // 监听窗口最大化/还原事件，通知渲染进程
-    if (mainWindow) {
-      mainWindow.on('maximize', () => {
-        mainWindow.webContents.send('window-maximized-state', true)
-      })
-      mainWindow.on('unmaximize', () => {
-        mainWindow.webContents.send('window-maximized-state', false)
-      })
-    }
-
-    // 设置项的 IPC 处理（简单存储到 userData 下）
-    ipcMain.handle('save-setting', (event, key, value) => {
-      try {
-        const settingsPath = path.join(app.getPath('userData'), 'knot-settings.json')
-        let settings = {}
-        if (fs.existsSync(settingsPath)) {
-          settings = JSON.parse(fs.readFileSync(settingsPath, 'utf8'))
-        }
-        settings[key] = value
-        fs.writeFileSync(settingsPath, JSON.stringify(settings, null, 2), 'utf8')
-        return true
-      } catch (err) {
-        console.error('保存设置失败:', err)
-        return false
-      }
-    })
-
-    // 重启应用
-    ipcMain.handle('restart-app', () => {
-      app.relaunch()
-      app.quit()
-    })
-
-    startBackend()
+    backendManager.start()
 
     setTimeout(() => {
       createWindow()
@@ -340,12 +136,12 @@ if (!gotTheLock) {
 }
 
 app.on('window-all-closed', () => {
-  stopBackend()
+  backendManager.stop()
   if (process.platform !== 'darwin') {
     app.quit()
   }
 })
 
 app.on('before-quit', () => {
-  stopBackend()
+  backendManager.stop()
 })
